Add whitelisted swap contract
Problem: there is a requirement to have a version of FA2 swaps contract
that allows a limited set of FA2 contracts.

Solution: add such a contract.

We use `simple_admin` component for administration functionality
necessary authorize whitelist updates.

Then add a component with whitelisting functionality, and in the same
file - the entire whitelisted swaps contract.