Pass big_map directly to `Set_allowed`
Problem: in `Set_allowed` entrypoint we currently accept a list of
addresses - for the sake of neater interface, and then convert it to
`big_map` form as in storage. But this increases costs of calling the
contract.

Solution: pass whitelist in `big_map` form to `Set_allowed`.

If this is not desired, feel free to revert this commit.