Move Haskell files at upper layer
Problem: currently all Haskell files lie in `test-hs` folder, and that
causes inconveniences.

For instance, there is a warning mentioning that we depend on `bin`
folder which outside of Haskell project. Also, I've put Haskell's `src`
to this `test-hs`, and this complicates the project structure.

Solution: move those files upper.

* Haskell project files go to `minter-contracts`;
* Haskell sources go from `test-hs/src` to `src-hs`;
* Haskell tests go from `test-hs/test` to `test-hs`.