Simplify allowlisted contracts organization
Problem: there is a need to evolve contracts in the direction of adding
other allowlist implementations. With the current organization this
turns out to be inconvenient.

Solution: the best we can end up for now is adding several base
allowlisting modules (one per implementation) that provide an
"interface" consisting of relevant types and functions. This is similar
to what admin modules do at the moment.

This commit does the mentioned reorganization. Allowlisting logic with
common helpers are put to respective modules. Each contract now contains
allowlisting logic for all the implementations of allowlisting.
`*allowlisted.mligo` files now just glue the respective base contract
module and allowlisting implementation.