Don't compile LIGO in jest test context
Closes #120

We now have a compilation process which compiles LIGO to TypeScript.
As such, the test helpers are now able to reference the compiled
TypeScript directly (removing the need to compile at test time).

An additional advantage of this approach is that we're testing the
same contract code consumers will be running.