Add configurable `Mutez` hedgehog generator
Problem: At the moment, Cleveland offers a Mutez generator that uses the
type's full range. However, in tests, you usually want the generated
values to be constrained to a certain range.

Solution: add a hedgehog generator for Mutez whose range can be
configured by the user. In order to be able to construct a mutez range,
we also have to have an instance of `Integral` for `Mutez` in scope, so
we added that instance as well.

This work has also been done in cleveland itself. So, the next time we
upgrade morley&friends, we should be able to delete the code added in
this commit.